backend/requirements.txt
asgiref==3.6.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.2
djoser==2.0.5
drf-extra-fields==3.4.1
psycopg2-binary==2.8.6
gunicorn==20.0.4
Pillow==9.4.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
webcolors==1.12